Add ReferenceTracker for storing many-to-many relationships in table storage
ReferenceTracker is optimized for cases where:
- one side of the relationship has a much higher cardinality than the other (e.g. more packages than certificates)
   - the high cardinality side is referred to as the "owner" and the other side is the "subject"
- the number of subjects per owned is relatively low (perhaps less than 100)
- edge deletion is rare (this path is not very optimized and a simple approach was taken)

For additional ease of use, data can be stored on the owner-to-subject edges.

Cleaning up unused subjects is done with a simple flow executed in isolation (singleton):
1. get all subjects that had any edges removed
2. check if each subject has any remaining owners
3. (perform some cleanup based on the orphan subjects)
4. clear the list of orphan candidates

This was designed for storing the many-to-many relationship between X.509 certificates (found in package signatures) and packages. Another commit will integrate this ReferenceTracker into a driver and into a singleton cleanup routine.