Change catalog index scan PK to be driver type not cursor name
When NuGet.Insights was first implemented, cursors were at the center of driver execution. Although that is still true to some extent, it's now reasonable to run a driver scan in other contexts (for example reprocessing specific packages without any cursor context). For this reason, a cursor update is more of an optional step in a driver execution.